+++ src/test/java/domain/height/HeightTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.height;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
