+++ src/test/java/domain/player/PlayersTest.java
@@ -1,4 +1,4 @@
-package domain.name;
+package domain.player;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
